requirements.txt
requests
python-dotenv
pymongo
faiss-cpu
sentence-transformers
py2neo
pandas
gseapy
